refactor: move renderer to test folder
remove rendererPipes since we don't use any other than default, even during testing
