+++ libs/frontend/abstract/application/src/renderer/renderer.dto.interface.ts
@@ -12,10 +12,6 @@ export interface IRendererDto {
    */
   containerNode: IComponentModel | IPageModel
   id: string
-  /**
-   * Allow override default for testing
-   */
-  renderPipe?: IRenderPipe
   rendererType: RendererType
   setSelectedNode?: IBuilderDomainService['setSelectedNode']
   urlSegments?: Record<string, string>
