+++ libs/frontend/application/renderer/src/index.ts
@@ -1,5 +1,3 @@
 export * from './renderer.application.service'
 export * from './RootRenderer'
-export * from './test/setup/renderer.test.factory'
-export * from './test/setup/root.test.store'
 export * from './utils'
