+++ libs/frontend/test/application/README.md
@@ -1,0 +1,7 @@
+# frontend-test-application
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test frontend-test-application` to execute the unit tests via [Vitest](https://vitest.dev/).
