+++ libs/frontend/test/application/jest.config.ts
@@ -1,0 +1,34 @@
+/* eslint-disable */
+export default {
+  displayName: 'frontend-test-application',
+  preset: '../../../../jest.preset.js',
+  transform: {
+    '^.+.[tj]sx?$': [
+      '@swc/jest',
+      {
+        jsc: {
+          parser: { syntax: 'typescript', tsx: true, decorators: true },
+          transform: {
+            decoratorMetadata: true,
+            react: { runtime: 'automatic' },
+          },
+        },
+      },
+    ],
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../../../coverage/libs/frontend/test/application',
+  reporters: [
+    'default',
+    [
+      'jest-junit',
+      {
+        outputName: 'frontend-test-application.xml',
+        reportTestSuiteErrors: true,
+        titleTemplate: '{classname} > {title}',
+        ancestorSeparator: ' › ',
+        classNameTemplate: '{displayName} > {filename}',
+      },
+    ],
+  ],
+}
