+++ libs/frontend/test/application/project.json
@@ -1,0 +1,27 @@
+{
+  "name": "frontend-test-application",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/frontend/test/application/src",
+  "projectType": "library",
+  "tags": ["scope:frontend"],
+  "targets": {
+    "lint": {
+      "executor": "@nx/eslint:lint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": [
+          "libs/frontend/test/application/**/*.{ts,tsx,js,jsx}"
+        ]
+      }
+    },
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/frontend/test/application/jest.config.ts",
+        "runInBand": true,
+        "reporters": ["default"]
+      }
+    }
+  }
+}
