+++ libs/frontend/test/application/src/renderer/runtime-component-props.spec.ts
