+++ libs/frontend/test/application/src/renderer/setup/index.ts
