+++ libs/frontend/test/application/src/renderer/setup/renderer.test.factory.ts
@@ -6,7 +6,6 @@ import type {
 import { RendererType } from '@codelab/frontend/abstract/application'
 import type { IPageModel } from '@codelab/frontend/abstract/domain'
 import { v4 } from 'uuid'
-import { PassThroughRenderPipe, renderPipeFactory } from '../../renderPipes'
 
 export const rendererFactory =
   (rendererService: IRendererService) => (dto: Partial<IRendererDto>) => {
@@ -16,8 +15,6 @@ export const rendererFactory =
       }) as IPageModel,
       id: dto.id ?? v4(),
       rendererType: dto.rendererType ?? RendererType.PageBuilder,
-      renderPipe: (dto.renderPipe ??
-        renderPipeFactory([PassThroughRenderPipe])) as IRenderPipe,
       urlSegments: dto.urlSegments ?? undefined,
     }
