+++ libs/frontend/test/application/src/renderer/typed-prop-tranformers.spec.ts
